backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.8.0
pydantic-settings==2.2.1
sqlalchemy==2.0.28
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pinecone-client==3.1.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-anthropic==0.1.4
langchain-community>=0.0.28,<0.1
openai==1.14.0
anthropic==0.19.1
tiktoken==0.6.0
pypdf==4.0.2
pdfminer.six==20221105
httpx==0.27.0
sse-starlette==2.0.0
slowapi==0.1.9
elevenlabs==2.3.0
pytest==8.0.2
pytest-asyncio==0.23.5
black==24.2.0
ruff==0.3.0
python-dotenv==1.0.1
# Queue system for async job processing
redis==5.0.1
rq==1.15.1
# File storage for production
boto3==1.34.34
# Enhanced text processing
python-magic==0.4.27
# Template engine for prompt system
jinja2==3.1.3

# FORCE CACHE BREAK - PostgreSQL drivers required 2025-06-16
